chore: finish counter example and README
