+++ examples/counter-postgres/config/default.toml
@@ -1,0 +1,13 @@
+[counter]
+entity-count   = 50
+evt-count      = 10_000
+snapshot-after = 100_000_000
+
+[evt-log]
+host          = "localhost"
+port          = 5432
+user          = "test"
+password      = "test"
+dbname        = "test"
+sslmode       = "prefer"
+poll-interval = "2s"
