--- examples/counter-postgres/proto/counter.proto
@@ -1,1 +1,0 @@
-../../../proto/counter.proto
