+++ examples/counter/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name          = "counter"
+description   = "eventsourced example base"
+version       = "0.1.0"
+publish       = false
+edition       = { workspace = true }
+authors       = { workspace = true }
+license       = { workspace = true }
+readme        = { workspace = true }
+homepage      = { workspace = true }
+repository    = { workspace = true }
+documentation = { workspace = true }
+
+[dependencies]
+eventsourced = { path = "../../eventsourced" }
+anyhow       = { workspace = true }
+bytes        = { workspace = true }
+serde        = { workspace = true }
+prost        = { workspace = true }
+tokio        = { workspace = true, features = [ "rt-multi-thread" ] }
+thiserror    = { workspace = true }
+tracing      = { workspace = true }
+uuid         = { workspace = true }
+
+[build-dependencies]
+anyhow      = { workspace = true }
+prost-build = { workspace = true }
+walkdir     = { workspace = true }
