+++ examples/counter/build.rs
