+++ examples/counter/proto/counter.proto
